+++ src/Server/Processes/Logger.php
@@ -7,10 +7,9 @@ use Innmind\Server\Control\Server\{
     Processes,
     Command,
     Signal,
-    Process,
     Process\Pid,
 };
-use Innmind\Immutable\Either;
+use Innmind\Immutable\Attempt;
 use Psr\Log\LoggerInterface;
 
 final class Logger implements Processes
@@ -27,7 +26,7 @@ final class Logger implements Processes
     }
 
     #[\Override]
-    public function execute(Command $command): Process
+    public function execute(Command $command): Attempt
     {
         $this->logger->info('About to execute a command', [
             'command' => $command->toString(),
@@ -41,7 +40,7 @@ final class Logger implements Processes
     }
 
     #[\Override]
-    public function kill(Pid $pid, Signal $signal): Either
+    public function kill(Pid $pid, Signal $signal): Attempt
     {
         $this->logger->info('About to kill a process', [
             'pid' => $pid->toInt(),