+++ src/Server/Processes/Remote.php
@@ -7,9 +7,7 @@ use Innmind\Server\Control\{
     Server\Processes,
     Server\Command,
     Server\Signal,
-    Server\Process,
     Server\Process\Pid,
-    ScriptFailed,
 };
 use Innmind\Url\Authority\{
     Host,
@@ -17,7 +15,7 @@ use Innmind\Url\Authority\{
     UserInformation\User,
 };
 use Innmind\Immutable\{
-    Either,
+    Attempt,
     SideEffect,
 };
 
@@ -47,7 +45,7 @@ final class Remote implements Processes
     }
 
     #[\Override]
-    public function execute(Command $command): Process
+    public function execute(Command $command): Attempt
     {
         /** @psalm-suppress ArgumentTypeCoercion Due psalm not understing that $bash cannot be empty */
         $command = $command
@@ -70,17 +68,14 @@ final class Remote implements Processes
     }
 
     #[\Override]
-    public function kill(Pid $pid, Signal $signal): Either
+    public function kill(Pid $pid, Signal $signal): Attempt
     {
-        $process = $this->execute(
-            $command = Command::foreground('kill')
-                ->withShortOption($signal->toString())
-                ->withArgument($pid->toString()),
-        );
-
-        return $process
-            ->wait()
-            ->map(static fn() => new SideEffect)
-            ->leftMap(static fn($e) => new ScriptFailed($command, $process, $e));
+        return $this
+            ->execute(
+                $command = Command::foreground('kill')
+                    ->withShortOption($signal->toString())
+                    ->withArgument($pid->toString()),
+            )
+            ->map(static fn() => SideEffect::identity());
     }
 }