+++ tests/Server/Processes/RemoteTest.php
@@ -21,7 +21,7 @@ use Innmind\Url\{
     Authority\UserInformation\User
 };
 use Innmind\Immutable\{
-    Either,
+    Attempt,
     SideEffect,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
@@ -57,7 +57,7 @@ class RemoteTest extends TestCase
             Process::class,
             $remote->execute(
                 Command::foreground('ls')->withShortOption('l'),
-            ),
+            )->unwrap(),
         );
     }
 
@@ -76,7 +76,7 @@ class RemoteTest extends TestCase
             Process::class,
             $remote->execute(
                 Command::foreground('ls')->withShortOption('l'),
-            ),
+            )->unwrap(),
         );
     }
 
@@ -96,7 +96,7 @@ class RemoteTest extends TestCase
                 Command::foreground('ls')
                     ->withShortOption('l')
                     ->withWorkingDirectory(Path::of('/tmp/foo')),
-            ),
+            )->unwrap(),
         );
     }
 
@@ -155,7 +155,7 @@ class RemoteTest extends TestCase
             ) {
             }
 
-            public function execute(Command $command): Process
+            public function execute(Command $command): Attempt
             {
                 $expected = \array_shift($this->commands);
                 $this->test->assertNotNull($expected);
@@ -167,7 +167,7 @@ class RemoteTest extends TestCase
                 return $this->processes->execute(Command::foreground('echo'));
             }
 
-            public function kill(Pid $pid, Signal $signal): Either
+            public function kill(Pid $pid, Signal $signal): Attempt
             {
             }
         };