+++ CHANGELOG.md
@@ -6,15 +6,25 @@
 
 - `Innmind\Server\Control\Server\Process::output()` now returns `Innmind\Immutable\Sequence<Innmind\Server\Control\Server\Process\Output\Chunk>`
 - `Innmind\Server\Control\Server\Process` is now a final class, old implementations are now declared internal
-- Requires `innmind/immutable:~5.0`
+- Requires `innmind/immutable:~5.12`
 - Requires `innmind/time-continuum:~4.1`
 - Requires `innmind/time-warp:~4.0`
 - Requires `innmind/filesystem:~8.0`
+- The following methods now return an `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`
+    - `Innmind\Server\Control\Server::reboot()`
+    - `Innmind\Server\Control\Server::shutdown()`
+    - `Innmind\Server\Control\Server::shutdown()`
+    - `Innmind\Server\Control\Server\Script::__invoke()`
+    - `Innmind\Server\Control\Server\Volumes::unmount()`
+    - `Innmind\Server\Control\Server\Volumes::unmount()`
+    - `Innmind\Server\Control\Server\Processes::kill()`
+- `Innmind\Server\Control\Server\Processes::execute()` now return an `Innmind\Immutable\Attempt<Innmind\Server\Control\Server\Process>`
 
 ### Removed
 
 - `Innmind\Server\Control\Server\Process\Output`
 - A process internal state is now longer logged
+- `Innmind\Server\Control\ScriptFailed`
 
 ### Fixed